test(turbomap): harden projection/zoom-lock + fix CI clippy gate
Follow-up to the projection + zoom-lock foundation: close the test gaps and
unbreak the clippy CI gate.

CI fix:
- `clippy --all-targets -D warnings` (the gated engine-crates lint) was red
  on main after the toolchain bump to 1.94.1, whose stricter
  `unnecessary_cast` flags two pre-existing `f64 as f64` casts in a camera
  test. Removed them; the full gate (core/scene/golden/engine/ffi/sim) is
  green again.

Test hardening:
- projection: an authoritative absolute anchor (no network/dep) — the
  WGS84 quarter meridian (10 001 965.729 m) pins the ellipsoid + arc series
  in metres, and an on-CM northing check pins the k0=0.9996 scale factor,
  the constants the round-trip/invariant tests leave free.
- ffi (gpu-tests): two host-surface integration tests — the zoom lock
  clamps a host-set camera both for a manual range and one auto-derived
  from a source that stops at z14; and a UTM33 coordinate projects to the
  same pixel as its WGS84 lat/lng (the no-drift contract, end-to-end).

Verified locally: workspace tests, the clippy gate, FFI Kotlin+Swift
binding generation (new UtmZone/ZoomRange/utmToLatLng/latLngToUtm/
setZoomBounds symbols present), and the GPU lanes via Lavapipe (golden +
engine + sim pass; the one `omt-bergen-3d` golden diff reproduces
identically on pristine main here — a Mesa-version pixel diff, not a code
regression).

+++ apps/turbomap/crates/turbomap-core/src/camera.rs
@@ -1041,14 +1041,14 @@ mod tests {
         // The centre lifts up (smaller y) by ~inset/2.
         assert!(inset_y < plain_y, "{inset_y} should sit above {plain_y}");
         assert!(
-            ((plain_y - inset_y) as f64 - inset / 2.0).abs() < 1.0,
+            ((plain_y - inset_y) - inset / 2.0).abs() < 1.0,
             "lift {} ≈ {}",
             plain_y - inset_y,
             inset / 2.0
         );
 
         // Round-trip still holds with the inset applied.
-        let back = inset_cam.pixel_to_world((512.0, inset_y as f64), viewport);
+        let back = inset_cam.pixel_to_world((512.0, inset_y), viewport);
         assert_world_close(back, world, 1e-3);
     }
